Replace some RaceEntry variables with helper functions, and made RaceEntry its own file.
